Bai tap 3

+++ src/main/java/j2ee_baitap/demo/model/Book.java
@@ -1,22 +1,22 @@
 package j2ee_baitap.demo.model;
 
 public class Book {
-    private int id;
+    private Long id; // Chuyển từ int sang Long để khớp với BookService
     private String title;
     private String author;
 
     // Constructors
     public Book() {}
 
-    public Book(int id, String title, String author) {
+    public Book(Long id, String title, String author) {
         this.id = id;
         this.title = title;
         this.author = author;
     }
 
     // Getters and Setters
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
